Basic Functionality
Added basic JS functionality to the Timer. It now counts down from 25 minutes and has a H3 which informs the user of what time that pomodoro will finish.

Still need to add the short breaks, long breaks & sets of Pomodoros to complete. Also added the time to the title section.